convert formal from regression to 3-class classification
- configs/vibes.json: formal now has classes 0=casual, 1=smart casual, 2=formal
- all 3 models: formal_head → Linear(*,3), MSELoss → CrossEntropyLoss
- src/clip_zero_shot.py: formal uses score_classification, removed score_formal
- src/data_loader.py: formal label dtype float32 → long
- src/evaluate.py: formal added to CLASS_DIMS, removed scatter plot, unified metrics

+++ src/data_loader.py
@@ -73,7 +73,7 @@ class FashionpediaDataset(Dataset):
         })
 
         labels = {
-            'formal':    torch.tensor(label_dict['formal'],     dtype=torch.float32),
+            'formal':    torch.tensor(label_dict['formal'],     dtype=torch.long),
             'season':    torch.tensor(label_dict['season'] - 1, dtype=torch.long),
             'gender':    torch.tensor(label_dict['gender'],     dtype=torch.long),
             'time':      torch.tensor(label_dict['time'],       dtype=torch.long),
